fixed gradient calculation rounding error

+++ LabSoftware VS2010/PolygonDrawer.cpp
@@ -74,8 +74,8 @@ namespace PolygonDrawer
 						output.push_back(p2);
 					} else
 					{
-						double gradient = (p2.y - p1.y)/(p2.x - p1.x);
-						int b = p1.y - gradient * p1.x;
+						double gradient = (p2.y - p1.y)/(double)(p2.x - p1.x);
+						double b = p1.y - gradient * p1.x;
      					VERTEX intersect = { (boundary - b) / gradient, boundary, p1.c };
 						output.push_back(intersect);
 						output.push_back(p2);
@@ -89,8 +89,8 @@ namespace PolygonDrawer
 						output.push_back(intersect);
 					} else
 					{
-						double gradient = (p2.y - p1.y)/(p2.x - p1.x);
-						int b = p2.y - gradient * p2.x;
+						double gradient = (p2.y - p1.y)/(double)(p2.x - p1.x);
+						double b = p2.y - gradient * p2.x;
 						VERTEX intersect = { (boundary - b) / gradient, boundary, p2.c };
 						output.push_back(intersect);
 					}					
@@ -110,8 +110,8 @@ namespace PolygonDrawer
 						output.push_back(p2);
 					} else
 					{
-						double gradient = (p2.y - p1.y)/(p2.x - p1.x);
-						int b = p1.y - gradient * p1.x;
+						double gradient = (p2.y - p1.y)/(double)(p2.x - p1.x);
+						double b = p1.y - gradient * p1.x;
 						VERTEX intersect = { (boundary - b) / gradient, boundary - 1, p1.c };
 						output.push_back(intersect);
 						output.push_back(p2);
@@ -125,8 +125,8 @@ namespace PolygonDrawer
 						output.push_back(intersect);
 					} else
 					{
-						double gradient = (p2.y - p1.y)/(p2.x - p1.x);
-						int b = p2.y - gradient * p2.x;
+						double gradient = (p2.y - p1.y)/(double)(p2.x - p1.x);
+						double b = p2.y - gradient * p2.x;
 						VERTEX intersect = { (boundary - b) / gradient, boundary - 1 , p2.c };
 						output.push_back(intersect);
 					}					
@@ -151,15 +151,15 @@ namespace PolygonDrawer
 			{
 				if (p1.x < boundary && p2.x >= boundary) //1st out, 2nd in
 				{
-					double gradient = (p2.y - p1.y)/(p2.x - p1.x);
-					int b = p1.y - gradient * p1.x;
+					double gradient = (p2.y - p1.y)/(double)(p2.x - p1.x);
+					double b = p1.y - gradient * p1.x;
 					VERTEX intersect = { boundary, gradient * boundary + b, p1.c };
 					output.push_back(intersect);
 					output.push_back(p2);
 				} else if (p2.x < boundary && p1.x >= boundary) //1st in, 2nd out
 				{
-					double gradient = (p2.y - p1.y)/(p2.x - p1.x);
-					int b = p2.y - gradient * p2.x;
+					double gradient = (p2.y - p1.y)/(double)(p2.x - p1.x);
+					double b = p2.y - gradient * p2.x;
 					VERTEX intersect = { boundary, gradient * boundary + b, p2.c };
 					output.push_back(intersect);
 				} else if (p1.x >= boundary && p2.x >= boundary)
@@ -170,15 +170,15 @@ namespace PolygonDrawer
 			{
 				if (p1.x >= boundary && p2.x < boundary) //1st out, 2nd in
 				{
-					double gradient = (p2.y - p1.y)/(p2.x - p1.x);
-					int b = p1.y - gradient * p1.x;
+					double gradient = (p2.y - p1.y)/(double)(p2.x - p1.x);
+					double b = p1.y - gradient * p1.x;
 					VERTEX intersect = { boundary - 1, gradient * (boundary - 1) + b, p1.c };
 					output.push_back(intersect);
 					output.push_back(p2);
 				} else if (p2.x >= boundary && p1.x < boundary) //1st in, 2nd out
 				{
-					double gradient = (p2.y - p1.y)/(p2.x - p1.x);
-					int b = p2.y - gradient * p2.x;
+					double gradient = (p2.y - p1.y)/(double)(p2.x - p1.x);
+					double b = p2.y - gradient * p2.x;
 					VERTEX intersect = { boundary - 1, gradient * (boundary - 1) + b, p2.c };
 					output.push_back(intersect);
 				} else if (p1.x < boundary && p2.x < boundary)
